requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
flatbuffers==23.1.21
fonttools==4.39.0
gast==0.4.0
google-auth==2.16.1
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
h5py==3.8.0
httplib2==0.21.0
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
jax==0.4.4
jedi==0.18.2
joblib==1.2.0
json5==0.9.11
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.12.0rc1
kiwisolver==1.4.4
labelImg==1.8.6
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.1
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip-tools==6.12.2
pipreqs==0.4.11
platformdirs==3.1.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.3
Pygments==2.14.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0rc0
tensorflow-intel==2.12.0rc0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.14.1
yarg==0.1.9